refs: remove the last remnants of core.preferSymlinkRefs
This step is for some time after Git 3.0.

Now it is almost N years since we removed the support for the
variable at Git 3.0 boundary, it is time to remove the warning about
this ancient variable and the behaviour change we went through with
it.

This concludes the journey to make sure we no longer create a
symbolic link to represent a symref, while still recognising
existing ones.

Signed-off-by: Junio C Hamano <gitster@pobox.com>