wip: do not say "worktree" when we mean a working tree
As "git help glossary" tells, a (non-bare) repository has its
primary working tree, with zero or more worktrees, each of which has
its own working tree.  The "worktree" is a wider concept that refers
to the combination of the working tree that house files and the Git
control files that are associated to the particular working tree.

Before the "worktree" was introduced, we used to call a working
tree as "worktree", and that remains in the variable names,
comments, and end-user facing messages.

Let's start fixing the end-user facing messages to avoid confusion.

There are places where the term "worktree" is used properly, namely
the "git branch" and its implementation that gives helper functions
to ensure that the same branch is not activated in more than one
worktrees.