refs_update_symref: atomically record overwritten ref
When updating a symref it's currently not possible to know for sure what
was the previous value that was overwritten. Make use of
ref_transaction's atomicity and record the previous value there. Add
a new variable to refs_update_symref's signature to be able to pass this
information back up to callers.

Signed-off-by: Bence Ferdinandy <bence@ferdinandy.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>