refspec.c: make valid_fetch_refspec() file-scope static
f2c6fda8 (refs: consolidate remote name validation, 2020-10-01)
moved valid_remote_name(), which receives the name the user gave to
"git remote add <name>" and checks if it is a name that we allow to
name a remote, to this file, even though it would have been a better
fit in remote.c file.  The function was the only caller of the
helper function valid_fetch_refspec() defined in this file, and in
the 4 years or so since then, we have not added any new callers to
the latter.

It could be argued that we should move valid_remote_name() to
remote.c someday, which would make it necessary for this function to
be public, but it is simpler to let the caller to stay there and
reduce the number of public functions by one.

Signed-off-by: Junio C Hamano <gitster@pobox.com>