bisect: optimize initial weighting of merges
"git bisect" is essentially about finding a commit, among N commits
that could be the real culprit that introduced a "change" we are
hunting for, that can reach about N/2 of them.  We call the number
of other commits that a commit can reach its "weight".

If you have a single-parent chain and know the weight of a commit
in the chain, the weight of a commit that has only one parent whose
weight W is known is easily computed (i.e. W+1, because it can reach
one commit more than its sole parent can).  But for a commit with
more than one parent, there is no efficient algorithm to count how
many of the other commits a commit can reach than actually counting,
which is quite costly.

So the central algorithm of bisecting consists of two phases: count
weight for all merge commits, and then propagate along single-parent
chains the weight of merge commits to their single-parent
descendants.  And in both phases, there is an early way out that
says "we found one commit whose weight is close enough to N/2, and
that is good enough".

In early stages of a big bisection session, where there are
thousands of commits in the range being bisected, we often hit these
early returns even from the initial phase to weigh merge commits.
To increase the chance of finding "good enough" merges before
wasting a lot of cycles, iterate over discovered merges (which are
ordered in the order of discovery) from the middle, as a commit in a
large graph in the middle has higher chance of being closer to the
middle.

We could further improve the algorithm by taking advantage of the
fact that parents of a commit that can reach way fewer than N/2
commits can never reach close to N/2 commits (simply because the
number of commits a parent of a commit can reach cannot be more than
the number of commits its child can reach) to cut down the cost of
this initial phase, but even without such optimization, the
heuristics to start from the middle and move outwards presented here
performs much better than the original algorithm.

Signed-off-by: Junio C Hamano <gitster@pobox.com>