libgitpub: move to separate contrib/ directory
For Cargo packaging purposes, it is better for any dependencies built with Git's
Makefile to live outside the libgit-{sys,rs} crates. This, plus some future
changes to the crates, will allow `cargo package` to treat all of the non-Rust
source as a separate tree apart from the Cargo-managed inputs.

Signed-off-by: Josh Steadmon <steadmon@google.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>