libgit-sys: add symlink to git repo root and build out of tree
Unlike `cargo build`, `cargo package` does not get access to the entire Git repo
containing a Rust crate. Instead, it prepares a directory starting from the
crate root (potentially excluding files, such as those not under version
control, or explicity excluded in the Cargo.toml file).

This means that the current method of building the libgit-sys crate does not
work with `cargo package`, as it tries to execute the Makefile from "../.."
relative to the crate root.

Fix this by adding a `git-src` symlink in the crate that points to the Git
repository root. `cargo package` will flatten this to a copy of the Git repo,
excluding non-version-controlled files, any explicitly-excluded files, and trees
that contain a Cargo.toml file (this prevents infinite recursion on the
symlink).

We can then execute the Makefile under the flattened git-src directory from our
build.rs script. However, this exposes a second problem; Cargo will check that
the build script does not add, delete, or modify any files in the source tree.
Without further changes, Cargo complains about the object files and other
generated files created during the build.

To avoid this problem, add a CARGO_OUT_DIR variable to the Makefile. When this
is set, object files and other generated files will be created there, rather
than in the main source tree. This change has only been applied to files created
as part of the libgitpub.a build, to avoid unnecessary churn in the Makefile.

Signed-off-by: Josh Steadmon <steadmon@google.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>